Moved testing code to separate file
